refactor : 제출물 상세 조회 책임 분리 및 중복 검증 로직 정리
