+++ src/test/java/com/opus/opus/restdocs/RestDocsTest.java
@@ -16,6 +16,7 @@ import com.opus.opus.modules.contest.application.ContestCategoryQueryService;
 import com.opus.opus.modules.contest.application.ContestCommandService;
 import com.opus.opus.modules.contest.application.ContestQueryService;
 import com.opus.opus.modules.contest.application.ContestSubmissionCommandService;
+import com.opus.opus.modules.contest.application.ContestSubmissionQueryService;
 import com.opus.opus.modules.contest.application.ContestTrackCommandService;
 import com.opus.opus.modules.contest.application.ContestTrackQueryService;
 import com.opus.opus.modules.member.api.MemberController;
@@ -111,6 +112,9 @@ public abstract class RestDocsTest extends ApiTestHelper {
 
     @MockitoBean
     protected ContestSubmissionCommandService contestSubmissionCommandService;
+
+    @MockitoBean
+    protected ContestSubmissionQueryService contestSubmissionQueryService;
 
     @MockitoBean
     protected ContestCategoryCommandService contestCategoryCommandService;
